requirements.txt
google-cloud-storage==1.16.1
earthengine-api
rasterio